Introduce gauge
The recorder has now a gauge method that generate two counters, one for the additions and one for the removals.

This allow the recording system to record a notion of 'volatility'.

E.g. even tough two consecutive measures may return the same value, this value may have change quite a bit in between. The 'adds' and 'removes' counter will expose that, and to retrieve the gauge, you just need to substract 'removes' from 'adds'.