Refactor Gauges
Gauges are now implemented as a simple number exported
under the `gauges` scope in the report.

Previously it was exported as two counters (adds/removes)
under the `counter` scope, which could trigger inconsistencies
when a gauge only increases (for instance) as the `removes`
counter won't be exported at the next report.